fix(utilities): rename urls file [KHCP-7705]
