+++ src/utilities/index.ts
@@ -1,3 +1,3 @@
 export { copyTextToClipboard } from './copyTextToClipboard'
 export { debounce } from './debounce'
-export { isValidUrl } from './helpers'
+export { isValidUrl } from './urls'
